Module massa_proto::google::api::http_rule

Nested message and enum types in HttpRule.

Structs

HttpRuleComment

Determines the URL pattern is matched by this rules. This pattern can be used with any of the {get|put|post|delete|patch} methods. A custom method can be defined using the ‘custom’ field.

Enums

Pattern

HACK: see docs in HttpRuleComment ignored in doctest pass

Functions

ScalarWrapper 🔒

ScalarWrapper 🔒

ScalarWrapper 🔒

ScalarWrapper 🔒

ScalarWrapper 🔒