Enum massa_proto::google::api::http_rule::Pattern

pub enum Pattern {
    Get(String),
    Put(String),
    Post(String),
    Delete(String),
    Patch(String),
    Custom(CustomHttpPattern),
}

HACK: see docs in HttpRuleComment ignored in doctest pass

Variants

Get(String)

Maps to HTTP GET. Used for listing and getting information about resources.

Put(String)

Maps to HTTP PUT. Used for replacing a resource.

Post(String)

Maps to HTTP POST. Used for creating a resource or performing an action.

Delete(String)

Maps to HTTP DELETE. Used for deleting a resource.

Patch(String)

Maps to HTTP PATCH. Used for updating a resource.

Custom(CustomHttpPattern)

The custom pattern is used for specifying an HTTP method that is not included in the pattern field, such as HEAD, or “*” to leave the HTTP method unspecified for this rule. The wild-card rule is useful for services that provide content to Web (HTML) clients.

Implementations

impl Pattern

pub fn encode<B>(&self, buf: &mut B)where B: BufMut,

Encodes the message to a buffer.

pub fn merge<B>( field: &mut Option<Pattern>, tag: u32, wire_type: WireType, buf: &mut B, ctx: DecodeContext ) -> Result<(), DecodeError>where B: Buf,

Decodes an instance of the message from a buffer, and merges it into self.

pub fn encoded_len(&self) -> usize

Returns the encoded length of the message without a length delimiter.