Module massa_proto::massa::api::v1

Version 1 of the Massa protos

Modules

massa_service_client

Generated client implementations.

massa_service_server

Generated server implementations.

send_blocks_response

Nested message and enum types in SendBlocksResponse.

send_endorsements_response

Nested message and enum types in SendEndorsementsResponse.

send_operations_response

Nested message and enum types in SendOperationsResponse.

Structs

Block

Block

BlockHeader

Block header

BlockParent

Block parent tuple

BlockResult

Holds Block response

BytesMapFieldEntry

BytesMapFieldEntry

CallSc

Calls an exported function from a stored smart contract

DatastoreEntriesQuery

DatastoreEntries Query

DatastoreEntry

DatastoreEntry

DatastoreEntryFilter

Endorsement

An endorsement, as sent in the network

EndorsementResult

Holds Endorsement response

ExecuteSc

Execute a smart contract

FilledBlock

Filled block

FilledOperationTuple

Filled Operation Tuple

GetBlocksBySlotsRequest

GetBlocksBySlotsRequest holds request for GetBlocksBySlots

GetBlocksBySlotsResponse

GetBlocksBySlotsResponse holds response from GetBlocksBySlots

GetDatastoreEntriesRequest

GetDatastoreEntriesRequest holds request from GetDatastoreEntries

GetDatastoreEntriesResponse

GetDatastoreEntriesResponse holds response from GetDatastoreEntries

GetNextBlockBestParentsRequest

GetNextBlockBestParentsRequest holds request for GetNextBlockBestParents

GetNextBlockBestParentsResponse

GetNextBlockBestParentsResponse holds response from GetNextBlockBestParents

GetSelectorDrawsRequest

GetSelectorDrawsRequest holds request from GetSelectorDraws

GetSelectorDrawsResponse

GetSelectorDrawsResponse holds response from GetSelectorDraws

GetTransactionsThroughputRequest

GetTransactionsThroughputRequest holds request for GetTransactionsThroughput

GetTransactionsThroughputResponse

GetTransactionsThroughputResponse holds response from GetTransactionsThroughput

GetVersionRequest

GetVersionRequest holds request from GetVersion

GetVersionResponse

GetVersionResponse holds response from GetVersion

IndexedSlot

When an address is drawn to create an endorsement it is selected for a specific index

Inner 🔒

NewBlocksHeadersRequest

NewBlocksHeadersRequest holds request for NewBlocksHeaders

NewBlocksHeadersResponse

NewBlocksHeadersResponse holds response from NewBlocksHeaders

NewBlocksRequest

NewBlocksRequest holds request for NewBlocks

NewBlocksResponse

NewBlocksResponse holds response from NewBlocks

NewFilledBlocksRequest

NewFilledBlocksRequest holds request for NewFilledBlocks

NewFilledBlocksResponse

NewFilledBlocksResponse holds response from NewFilledBlocks

NewOperationsFilter

NewOperations Filter

NewOperationsQuery

NewOperations Query

NewOperationsRequest

NewOperationsRequest holds request for NewOperations

NewOperationsResponse

NewOperationsResponse holds response from NewOperations

Operation

The operation as sent in the network

OperationResult

Holds Operation response

OperationType

Type specific operation content

RollBuy

The sender buys roll_count rolls. Roll price is defined in configuration

RollSell

The sender sells roll_count rolls. Roll price is defined in configuration

ScalarWrapper 🔒

ScalarWrapper 🔒

ScalarWrapper 🔒

ScalarWrapper 🔒

ScalarWrapper 🔒

ScalarWrapper 🔒

SecureShare

Packages a type such that it can be securely sent and received in a trust-free network

SelectorDraws

Selector draws

SelectorDrawsFilter

SelectorDraws Filter

SelectorDrawsQuery

SelectorDraws Query

SendBlocksRequest

SendBlocksRequest holds parameters to SendBlocks

SendBlocksResponse

SendBlocksResponse holds response from SendBlocks

SendEndorsementsRequest

SendEndorsementsRequest holds parameters to SendEndorsements

SendEndorsementsResponse

SendEndorsementsResponse holds response from SendEndorsements

SendOperationsRequest

SendOperationsRequest holds parameters to SendOperations

SendOperationsResponse

SendOperationsResponse holds response from SendOperations

SignedBlock

Signed block

SignedBlockHeader

Signed block header

SignedEndorsement

Signed endorsement

SignedOperation

Signed operation

Slot

A point in time where a block is expected

Transaction

Transfer coins from sender to recipient

TransactionsThroughputRequest

TransactionsThroughputRequest holds request for TransactionsThroughput

TransactionsThroughputResponse

TransactionsThroughputResponse holds response from TransactionsThroughput

Enums

OpType

Operation type enum

Constants

FILE_DESCRIPTOR_SET

Compiled file descriptor set for the Massa protos

STRUCT_NAME 🔒

STRUCT_NAME 🔒

STRUCT_NAME 🔒

STRUCT_NAME 🔒

STRUCT_NAME 🔒

STRUCT_NAME 🔒

STRUCT_NAME 🔒

STRUCT_NAME 🔒

STRUCT_NAME 🔒

STRUCT_NAME 🔒

STRUCT_NAME 🔒

STRUCT_NAME 🔒

STRUCT_NAME 🔒

STRUCT_NAME 🔒

STRUCT_NAME 🔒

STRUCT_NAME 🔒

STRUCT_NAME 🔒

STRUCT_NAME 🔒

STRUCT_NAME 🔒

STRUCT_NAME 🔒

STRUCT_NAME 🔒

STRUCT_NAME 🔒

STRUCT_NAME 🔒

STRUCT_NAME 🔒

STRUCT_NAME 🔒

STRUCT_NAME 🔒

STRUCT_NAME 🔒

STRUCT_NAME 🔒

STRUCT_NAME 🔒

STRUCT_NAME 🔒

STRUCT_NAME 🔒

STRUCT_NAME 🔒

STRUCT_NAME 🔒

STRUCT_NAME 🔒

STRUCT_NAME 🔒

STRUCT_NAME 🔒

STRUCT_NAME 🔒

STRUCT_NAME 🔒

STRUCT_NAME 🔒

STRUCT_NAME 🔒

STRUCT_NAME 🔒

STRUCT_NAME 🔒

STRUCT_NAME 🔒

STRUCT_NAME 🔒

STRUCT_NAME 🔒

STRUCT_NAME 🔒

STRUCT_NAME 🔒

STRUCT_NAME 🔒

STRUCT_NAME 🔒

STRUCT_NAME 🔒

STRUCT_NAME 🔒

STRUCT_NAME 🔒

STRUCT_NAME 🔒

STRUCT_NAME 🔒

STRUCT_NAME 🔒

STRUCT_NAME 🔒

STRUCT_NAME 🔒

STRUCT_NAME 🔒

STRUCT_NAME 🔒

STRUCT_NAME 🔒

Functions

Inner 🔒

Inner 🔒

Inner 🔒

Inner 🔒

Inner 🔒

ScalarWrapper 🔒

ScalarWrapper 🔒

ScalarWrapper 🔒

ScalarWrapper 🔒

ScalarWrapper 🔒

ScalarWrapper 🔒

ScalarWrapper 🔒

ScalarWrapper 🔒

ScalarWrapper 🔒

ScalarWrapper 🔒

ScalarWrapper 🔒

ScalarWrapper 🔒

ScalarWrapper 🔒

ScalarWrapper 🔒

ScalarWrapper 🔒

ScalarWrapper 🔒

ScalarWrapper 🔒

ScalarWrapper 🔒

ScalarWrapper 🔒

ScalarWrapper 🔒

ScalarWrapper 🔒

ScalarWrapper 🔒

ScalarWrapper 🔒

ScalarWrapper 🔒

ScalarWrapper 🔒

ScalarWrapper 🔒

ScalarWrapper 🔒

ScalarWrapper 🔒

ScalarWrapper 🔒

ScalarWrapper 🔒

ScalarWrapper 🔒

ScalarWrapper 🔒

ScalarWrapper 🔒

ScalarWrapper 🔒

ScalarWrapper 🔒

ScalarWrapper 🔒

ScalarWrapper 🔒

ScalarWrapper 🔒

ScalarWrapper 🔒

ScalarWrapper 🔒

ScalarWrapper 🔒

ScalarWrapper 🔒

ScalarWrapper 🔒

ScalarWrapper 🔒

ScalarWrapper 🔒

ScalarWrapper 🔒

ScalarWrapper 🔒

ScalarWrapper 🔒

ScalarWrapper 🔒

ScalarWrapper 🔒

ScalarWrapper 🔒

ScalarWrapper 🔒

ScalarWrapper 🔒

ScalarWrapper 🔒

ScalarWrapper 🔒

ScalarWrapper 🔒

ScalarWrapper 🔒

ScalarWrapper 🔒

ScalarWrapper 🔒

ScalarWrapper 🔒

ScalarWrapper 🔒

ScalarWrapper 🔒

ScalarWrapper 🔒

ScalarWrapper 🔒

ScalarWrapper 🔒

ScalarWrapper 🔒

ScalarWrapper 🔒

ScalarWrapper 🔒

ScalarWrapper 🔒

ScalarWrapper 🔒

ScalarWrapper 🔒

ScalarWrapper 🔒

ScalarWrapper 🔒

ScalarWrapper 🔒

ScalarWrapper 🔒

ScalarWrapper 🔒

ScalarWrapper 🔒

ScalarWrapper 🔒

ScalarWrapper 🔒

ScalarWrapper 🔒

ScalarWrapper 🔒

ScalarWrapper 🔒

ScalarWrapper 🔒

ScalarWrapper 🔒