Struct massa_proto::massa::api::v1::massa_service_client::MassaServiceClient

pub struct MassaServiceClient<T> {
    inner: Grpc<T>,
}

Massa gRPC service

Fields
inner: Grpc<T>
Implementations

impl MassaServiceClient<Channel>

pub async fn connect<D>(dst: D) -> Result<Self, Error>where D: TryInto<Endpoint>, D::Error: Into<StdError>,

Attempt to create a new client by connecting to a given endpoint.

impl<T> MassaServiceClient<T>where T: GrpcService<BoxBody>, T::Error: Into<StdError>, T::ResponseBody: Body<Data = Bytes> + Send + 'static, <T::ResponseBody as Body>::Error: Into<StdError> + Send,

pub fn new(inner: T) -> Self

pub fn with_origin(inner: T, origin: Uri) -> Self

pub fn with_interceptor<F>( inner: T, interceptor: F ) -> MassaServiceClient<InterceptedService<T, F>>where F: Interceptor, T::ResponseBody: Default, T: Service<Request<BoxBody>, Response = Response<<T as GrpcService<BoxBody>>::ResponseBody>>, <T as Service<Request<BoxBody>>>::Error: Into<StdError> + Send + Sync,

pub fn send_compressed(self, encoding: CompressionEncoding) -> Self

Compress requests with the given encoding.

This requires the server to support it otherwise it might respond with an error.

pub fn accept_compressed(self, encoding: CompressionEncoding) -> Self

Enable decompressing responses.

pub fn max_decoding_message_size(self, limit: usize) -> Self

Limits the maximum size of a decoded message.

Default: 4MB

pub fn max_encoding_message_size(self, limit: usize) -> Self

Limits the maximum size of an encoded message.

Default: usize::MAX

pub async fn get_blocks( &mut self, request: impl IntoRequest<GetBlocksRequest> ) -> Result<Response<GetBlocksResponse>, Status>

Get blocks by ids

pub async fn get_blocks_by_slots( &mut self, request: impl IntoRequest<GetBlocksBySlotsRequest> ) -> Result<Response<GetBlocksBySlotsResponse>, Status>

Get blocks by slots

pub async fn get_datastore_entries( &mut self, request: impl IntoRequest<GetDatastoreEntriesRequest> ) -> Result<Response<GetDatastoreEntriesResponse>, Status>

Get datastore entries

pub async fn get_largest_stakers( &mut self, request: impl IntoRequest<GetLargestStakersRequest> ) -> Result<Response<GetLargestStakersResponse>, Status>

Get largest stakers

pub async fn get_next_block_best_parents( &mut self, request: impl IntoRequest<GetNextBlockBestParentsRequest> ) -> Result<Response<GetNextBlockBestParentsResponse>, Status>

Get next block best parents

pub async fn get_operations( &mut self, request: impl IntoRequest<GetOperationsRequest> ) -> Result<Response<GetOperationsResponse>, Status>

Get operations

pub async fn get_sc_execution_events( &mut self, request: impl IntoRequest<GetScExecutionEventsRequest> ) -> Result<Response<GetScExecutionEventsResponse>, Status>

Get smart contracts execution events

pub async fn get_selector_draws( &mut self, request: impl IntoRequest<GetSelectorDrawsRequest> ) -> Result<Response<GetSelectorDrawsResponse>, Status>

Get selector draws

pub async fn get_transactions_throughput( &mut self, request: impl IntoRequest<GetTransactionsThroughputRequest> ) -> Result<Response<GetTransactionsThroughputResponse>, Status>

Get transactions throughput

pub async fn get_version( &mut self, request: impl IntoRequest<GetVersionRequest> ) -> Result<Response<GetVersionResponse>, Status>

Get node version

pub async fn new_blocks( &mut self, request: impl IntoStreamingRequest<Message = NewBlocksRequest> ) -> Result<Response<Streaming<NewBlocksResponse>>, Status>

New received and produced blocks

pub async fn new_blocks_headers( &mut self, request: impl IntoStreamingRequest<Message = NewBlocksHeadersRequest> ) -> Result<Response<Streaming<NewBlocksHeadersResponse>>, Status>

New received and produced blocks headers

pub async fn new_endorsements( &mut self, request: impl IntoStreamingRequest<Message = NewEndorsementsRequest> ) -> Result<Response<Streaming<NewEndorsementsResponse>>, Status>

New received and produced endorsements

pub async fn new_filled_blocks( &mut self, request: impl IntoStreamingRequest<Message = NewFilledBlocksRequest> ) -> Result<Response<Streaming<NewFilledBlocksResponse>>, Status>

New received and produced blocks with operations

pub async fn new_operations( &mut self, request: impl IntoStreamingRequest<Message = NewOperationsRequest> ) -> Result<Response<Streaming<NewOperationsResponse>>, Status>

New received and produced operations

pub async fn new_slot_execution_outputs( &mut self, request: impl IntoStreamingRequest<Message = NewSlotExecutionOutputsRequest> ) -> Result<Response<Streaming<NewSlotExecutionOutputsResponse>>, Status>

New received and produced smart contract execution events

pub async fn send_blocks( &mut self, request: impl IntoStreamingRequest<Message = SendBlocksRequest> ) -> Result<Response<Streaming<SendBlocksResponse>>, Status>

Send blocks

pub async fn send_endorsements( &mut self, request: impl IntoStreamingRequest<Message = SendEndorsementsRequest> ) -> Result<Response<Streaming<SendEndorsementsResponse>>, Status>

Send endorsements

pub async fn send_operations( &mut self, request: impl IntoStreamingRequest<Message = SendOperationsRequest> ) -> Result<Response<Streaming<SendOperationsResponse>>, Status>

Send operations

pub async fn transactions_throughput( &mut self, request: impl IntoStreamingRequest<Message = TransactionsThroughputRequest> ) -> Result<Response<Streaming<TransactionsThroughputResponse>>, Status>

Transactions throughput