Module massa_proto::massa::api::v1

Version 1 of the Massa protos

Modules

grpc_client

Generated client implementations.

grpc_server

Generated server implementations.

send_blocks_stream_response

Nested message and enum types in SendBlocksStreamResponse.

send_endorsements_stream_response

Nested message and enum types in SendEndorsementsStreamResponse.

send_operations_stream_response

Nested message and enum types in SendOperationsStreamResponse.

Structs

Block

Block

BlockHeader

Block header

BlockParent

Block parent tuple

BlockResult

Holds Block response

BytesMapFieldEntry

BytesMapFieldEntry

CallSc

Calls an exported function from a stored smart contract

DatastoreEntriesQuery

DatastoreEntries Query

DatastoreEntry

DatastoreEntry

DatastoreEntryFilter

Endorsement

An endorsement, as sent in the network

EndorsementResult

Holds Endorsement response

ExecuteSc

Execute a smart contract

FilledBlock

Filled block

FilledOperationTuple

Filled Operation Tuple

GetBlocksBySlotsRequest

GetBlocksBySlotsRequest holds request for GetBlocksBySlots

GetBlocksBySlotsResponse

GetBlocksBySlotsResponse holds response from GetBlocksBySlots

GetDatastoreEntriesRequest

GetDatastoreEntriesRequest holds request from GetDatastoreEntries

GetDatastoreEntriesResponse

GetDatastoreEntriesResponse holds response from GetDatastoreEntries

GetNextBlockBestParentsRequest

GetNextBlockBestParentsRequest holds request for GetNextBlockBestParents

GetNextBlockBestParentsResponse

GetNextBlockBestParentsResponse holds response from GetNextBlockBestParents

GetSelectorDrawsRequest

GetSelectorDrawsRequest holds request from GetSelectorDraws

GetSelectorDrawsResponse

GetSelectorDrawsResponse holds response from GetSelectorDraws

GetTransactionsThroughputRequest

GetTransactionsThroughputRequest holds request for GetTransactionsThroughput

GetTransactionsThroughputResponse

GetTransactionsThroughputResponse holds response from GetTransactionsThroughput

GetVersionRequest

GetVersionRequest holds request from GetVersion

GetVersionResponse

GetVersionResponse holds response from GetVersion

IndexedSlot

When an address is drawn to create an endorsement it is selected for a specific index

NewBlocksHeadersStreamRequest

NewBlocksHeadersStreamRequest holds request for NewBlocksHeadersStream

NewBlocksHeadersStreamResponse

NewBlocksHeadersStreamResponse holds response from NewBlocksHeadersStream

NewBlocksStreamRequest

NewBlocksStreamRequest holds request for NewBlocksStream

NewBlocksStreamResponse

NewBlocksStreamResponse holds response from NewBlocksStream

NewFilledBlocksStreamRequest

NewFilledBlocksStreamRequest holds request for NewFilledBlocksStream

NewFilledBlocksStreamResponse

NewFilledBlocksStreamResponse holds response from NewFilledBlocksStream

NewOperationsStreamFilter

NewOperationsStream Filter

NewOperationsStreamQuery

NewOperationsStream Query

NewOperationsStreamRequest

NewOperationsStreamRequest holds request for NewOperationsStream

NewOperationsStreamResponse

NewOperationsStreamResponse holds response from NewOperationsStream

Operation

The operation as sent in the network

OperationResult

Holds Operation response

OperationType

Type specific operation content

RollBuy

The sender buys roll_count rolls. Roll price is defined in configuration

RollSell

The sender sells roll_count rolls. Roll price is defined in configuration

SecureShare

Packages a type such that it can be securely sent and received in a trust-free network

SelectorDraws

Selector draws

SelectorDrawsFilter

SelectorDraws Filter

SelectorDrawsQuery

SelectorDraws Query

SendBlocksStreamRequest

SendBlocksStreamRequest holds parameters to SendBlocks

SendBlocksStreamResponse

SendBlocksStreamResponse holds response from SendBlocks

SendEndorsementsStreamRequest

SendEndorsementsStreamRequest holds parameters to SendEndorsements

SendEndorsementsStreamResponse

SendEndorsementsStreamResponse holds response from SendEndorsements

SendOperationsStreamRequest

SendOperationsStreamRequest holds parameters to SendOperations

SendOperationsStreamResponse

SendOperationsStreamResponse holds response from SendOperations

SignedBlock

Signed block

SignedBlockHeader

Signed block header

SignedEndorsement

Signed endorsement

SignedOperation

Signed operation

Slot

A point in time where a block is expected

Transaction

Transfer coins from sender to recipient

TransactionsThroughputStreamRequest

TransactionsThroughputStreamRequest holds request for TransactionsThroughputStream

Enums

OperationTypeEnum

Operation type enum

Constants

FILE_DESCRIPTOR_SET

Compiled file descriptor set for the Massa protos